nova-act/usecases/adaptive-ui-testing-with-nova-act/nova-act-tests/requirements.txt
nova-act[cli]>=1.0.0
pytest>=7.0.0
pytest-html>=3.0.0
python-dotenv>=1.0.0
